+++ libnavui-speedlimit/src/main/java/com/mapbox/navigation/ui/speedlimit/model/UpdateSpeedLimitError.kt
@@ -6,6 +6,13 @@ package com.mapbox.navigation.ui.speedlimit.model
  * @param errorMessage an error message
  * @param throwable an optional throwable value expressing the error
  */
+@Deprecated(
+    message = "This data object is being deprecated.",
+    replaceWith = ReplaceWith(
+        expression = "SpeedInfoError",
+        imports = arrayOf("SpeedInfoError")
+    )
+)
 class UpdateSpeedLimitError internal constructor(
     val errorMessage: String,
     val throwable: Throwable?